Amiibo efficiency change and invalid user input correction
Amiibo method "getNumberOfMatchesAgainstAmiiboSum" was calling the "getNumberOfMatchesAgainstAmiibo" method for a return value rather than just using the private "numberOfMatchesAgainstAmiibo" variable directly. There was also an infinite loop issue when scanning for an integer from the user that would cause infinite error messages to appear that has been fixed. There was also a typo in one of the "importSpreadsheetData" method's comments that has been corrected.